fix: remove clever workaround for /follow
Users have been issues /follow to get the most recent entry to show up and getting mad that it skips other entries. The follow command posted the most recent entry in all channels that it made sense as a clever work around.
Instead we need to limit the how it works compeltely.